requirements.txt
fastapi
uvicorn[standard]
httpx
fastapi-mcp

anthropic
Pillow
opencv-python
gTTS
IPython
playsound==1.2.2
PyObjC 